requirements.txt
fastapi==0.115.12
grpcio==1.71.0
grpcio-tools==1.71.0
requests==2.32.3
streamlit==1.45.0
uvicorn==0.34.2
docopt==0.6.2 
num2words==0.5.14
soundfile==0.13.1
streamlit-lottie==0.0.5
torch==2.7.0
datasets==3.5.1
torchaudio==2.7.0
transformers==4.51.3
sentencepiece==0.2.0